Use new API.

+++ fixers/debian-tests-control-autodep8-is-obsolete.py
@@ -2,7 +2,7 @@
 
 import os
 import sys
-from lintian_brush.fixer import fixed_lintian_tag, report_result
+from lintian_brush.fixer import LintianIssue, report_result
 
 OLD_PATH = 'debian/tests/control.autodep8'
 NEW_PATH = 'debian/tests/control'
@@ -11,20 +11,28 @@ if not os.path.exists(OLD_PATH):
     sys.exit(0)
 
 
-fixed_lintian_tag(
+issue = LintianIssue(
     'source', 'debian-tests-control-autodep8-is-obsolete',
     info=OLD_PATH)
 
+if not issue.should_fix():
+    sys.exit(0)
+
 if not os.path.exists(NEW_PATH):
     os.rename(OLD_PATH, 'debian/tests/control')
+    issue.report_fixed()
     report_result("Rename obsolete path %s to %s." % (OLD_PATH, NEW_PATH))
 else:
+    merge_issue = LintianIssue(
+        'source', 'debian-tests-control-and-control-autodep8',
+        info='%s %s' % (OLD_PATH, NEW_PATH))
+    if not merge_issue.should_fix():
+        sys.exit(0)
+    merge_issue.report_fixed()
+    issue.report_fixed()
     with open(NEW_PATH, 'ab') as outf:
         outf.write(b'\n')
         with open(OLD_PATH, 'rb') as inf:
             outf.writelines(inf.readlines())
     os.unlink(OLD_PATH)
-    fixed_lintian_tag(
-        'source', 'debian-tests-control-and-control-autodep8',
-        info='%s %s' % (OLD_PATH, NEW_PATH))
     report_result("Merge %s into %s." % (OLD_PATH, NEW_PATH))
