+++ fixers/empty-debian-tests-control.py
@@ -1,15 +1,17 @@
 #!/usr/bin/python3
 
 import os
-from lintian_brush.fixer import report_result, fixed_lintian_tag
+from lintian_brush.fixer import report_result, LintianIssue
 
 try:
     with open('debian/tests/control', 'r') as f:
         if f.read().strip() == '':
-            os.unlink('debian/tests/control')
-            if not os.listdir('debian/tests'):
-                os.rmdir('debian/tests')
-            fixed_lintian_tag('source', 'empty-debian-tests-control')
+            issue = LintianIssue('source', 'empty-debian-tests-control')
+            if issue.should_fix():
+                os.unlink('debian/tests/control')
+                if not os.listdir('debian/tests'):
+                    os.rmdir('debian/tests')
+                issue.report_fixed()
 except (FileNotFoundError, NotADirectoryError):
     pass
